# [Performance] Avoid $reflector->reflectAllClasses() in every worker on DynamicSourceLocatorProvider