--- a/phpstan.neon
+++ b/phpstan.neon
@@ -432,3 +432,7 @@ parameters:
             message: '#Parameters should use "array" types as the only types passed to this method#'
             path: src/Util/NodePrinter.php
         - '#Cognitive complexity for "Rector\\TypeDeclaration\\NodeAnalyzer\\CallerParamMatcher\:\:matchCallParamType\(\)" is 12, keep it under 11#'
+
+        -
+            message: '#Use required typed property over of nullable array property#'
+            path: src/NodeTypeResolver/Reflection/BetterReflection/SourceLocatorProvider/DynamicSourceLocatorProvider.php