--- a/src/NodeTypeResolver/Reflection/BetterReflection/SourceLocatorProvider/DynamicSourceLocatorProvider.php
+++ b/src/NodeTypeResolver/Reflection/BetterReflection/SourceLocatorProvider/DynamicSourceLocatorProvider.php
@@ -162,6 +162,11 @@ final class DynamicSourceLocatorProvider implements ResetableInterface
             return;
         }
 
+        $this->reflectAndGetAllClasses($aggregateSourceLocator);
+    }
+
+    private function reflectAndGetAllClasses(AggregateSourceLocator $aggregateSourceLocator): void
+    {
         $reflector = new DefaultReflector($aggregateSourceLocator);
         $this->classNames = [];