This should improve performance to avoid `$reflector->reflectAllClasses()` called on every worker on parallel process which scan every single files to get Class like.

Closes https://github.com/rectorphp/rector/issues/8637
Closes https://github.com/rectorphp/rector/issues/8638

/cc @mfn @dorrogeray 

**Before**

```bash
➜  rector-src git:(main) time bin/rector process --clear-cache                                                                             
 2180/2180 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  378.25s user 15.69s system 706% cpu 55.720 total
```

**After**

```bash
➜  rector-src git:(avoid-reflect) time bin/rector process --clear-cache
 2180/2180 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  227.14s user 9.48s system 621% cpu 38.062 total
```